requirements.txt
azure-cognitiveservices-speech
azure-ai-language-questionanswering